+++ poem/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# Unreleased
+
+- Panic when there are duplicates in the routing table. [#126](https://github.com/poem-web/poem/issues/126)
+
 # [1.2.4] 2021-12-17
 
 - Rename `EndpointExt::inspect_error` to `EndpointExt::inspect_all_error`.